requirements.txt
streamlit==1.35.0
pandas==2.2.2
sentence-transformers==2.6.1
faiss-cpu==1.7.4
numpy==1.24.4
openai==0.28.0
